extract ParquetFileReaderFactory test to integration tests
