+++ datafusion/core/src/datasource/file_format/parquet.rs
@@ -354,7 +354,11 @@ fn summarize_min_max(
     }
 }
 
-pub(crate) async fn fetch_parquet_metadata(
+/// Fetches parquet metadata from ObjectStore for given object
+///
+/// This component is a subject to **change** in near future and is exposed for low level integrations
+/// through [ParquetFileReaderFactory].
+pub async fn fetch_parquet_metadata(
     store: &dyn ObjectStore,
     meta: &ObjectMeta,
     size_hint: Option<usize>,
@@ -552,47 +556,6 @@ pub(crate) mod test_util {
 
         let meta: Vec<_> = files.iter().map(local_unpartitioned_file).collect();
         Ok((meta, files))
-    }
-
-    pub async fn store_parquet_in_memory(
-        batches: Vec<RecordBatch>,
-    ) -> (Arc<dyn ObjectStore>, Vec<ObjectMeta>) {
-        let in_memory = InMemory::new();
-
-        let parquet_batches: Vec<(ObjectMeta, Bytes)> = batches
-            .into_iter()
-            .enumerate()
-            .map(|(offset, batch)| {
-                let mut buf = Vec::<u8>::with_capacity(32 * 1024);
-                let mut output = Cursor::new(&mut buf);
-
-                let mut writer = ArrowWriter::try_new(&mut output, batch.schema(), None)
-                    .expect("creating writer");
-
-                writer.write(&batch).expect("Writing batch");
-                writer.close().unwrap();
-
-                let meta = ObjectMeta {
-                    location: Path::parse(format!("file-{offset}.parquet"))
-                        .expect("creating path"),
-                    last_modified: chrono::DateTime::from(SystemTime::now()),
-                    size: buf.len(),
-                };
-
-                (meta, Bytes::from(buf))
-            })
-            .collect();
-
-        let mut objects = Vec::with_capacity(parquet_batches.len());
-        for (meta, bytes) in parquet_batches {
-            in_memory
-                .put(&meta.location, bytes)
-                .await
-                .expect("put parquet file into in memory object store");
-            objects.push(meta);
-        }
-
-        (Arc::new(in_memory), objects)
     }
 }
