+++ datafusion/core/src/physical_plan/file_format/parquet.rs
@@ -86,17 +86,6 @@ pub struct ParquetExec {
     parquet_file_reader_factory: Option<Arc<dyn ParquetFileReaderFactory>>,
 }
 
-/// Stores metrics about the parquet execution for a particular parquet file
-#[derive(Debug, Clone)]
-struct ParquetFileMetrics {
-    /// Number of times the predicate could not be evaluated
-    pub predicate_evaluation_errors: metrics::Count,
-    /// Number of row groups pruned using
-    pub row_groups_pruned: metrics::Count,
-    /// Total number of bytes scanned
-    pub bytes_scanned: metrics::Count,
-}
-
 impl ParquetExec {
     /// Create a new Parquet reader execution plan provided file list and schema.
     pub fn new(
@@ -162,6 +151,20 @@ impl ParquetExec {
         self.parquet_file_reader_factory = Some(parquet_file_reader_factory);
         self
     }
+}
+
+/// Stores metrics about the parquet execution for a particular parquet file.
+///
+/// This component is a subject to **change** in near future and is exposed for low level integrations
+/// through [ParquetFileReaderFactory].
+#[derive(Debug, Clone)]
+pub struct ParquetFileMetrics {
+    /// Number of times the predicate could not be evaluated
+    pub predicate_evaluation_errors: metrics::Count,
+    /// Number of row groups pruned using
+    pub row_groups_pruned: metrics::Count,
+    /// Total number of bytes scanned
+    pub bytes_scanned: metrics::Count,
 }
 
 impl ParquetFileMetrics {
@@ -412,8 +415,8 @@ impl DefaultParquetFileReaderFactory {
 struct ParquetFileReader {
     store: Arc<dyn ObjectStore>,
     meta: ObjectMeta,
-    metadata_size_hint: Option<usize>,
     metrics: ParquetFileMetrics,
+    metadata_size_hint: Option<usize>,
 }
 
 impl AsyncFileReader for ParquetFileReader {
@@ -858,105 +861,6 @@ mod tests {
     }
 
     const EXPECTED_USER_DEFINED_METADATA: &str = "some-user-defined-metadata";
-
-    #[tokio::test]
-    async fn route_data_access_ops_to_parquet_file_reader_factory() {
-        let c1: ArrayRef =
-            Arc::new(StringArray::from(vec![Some("Foo"), None, Some("bar")]));
-        let c2: ArrayRef = Arc::new(Int64Array::from(vec![Some(1), Some(2), None]));
-        let c3: ArrayRef = Arc::new(Int8Array::from(vec![Some(10), Some(20), None]));
-
-        let batch = create_batch(vec![
-            ("c1", c1.clone()),
-            ("c2", c2.clone()),
-            ("c3", c3.clone()),
-        ]);
-
-        let file_schema = batch.schema().clone();
-        let (in_memory_object_store, parquet_files_meta) =
-            store_parquet_in_memory(vec![batch]).await;
-        let file_groups = parquet_files_meta
-            .into_iter()
-            .map(|meta| PartitionedFile {
-                object_meta: meta,
-                partition_values: vec![],
-                range: None,
-                extensions: Some(Arc::new(String::from(EXPECTED_USER_DEFINED_METADATA))),
-            })
-            .collect();
-
-        // prepare the scan
-        let parquet_exec = ParquetExec::new(
-            FileScanConfig {
-                // just any url that doesn't point to in memory object store
-                object_store_url: ObjectStoreUrl::local_filesystem(),
-                file_groups: vec![file_groups],
-                file_schema,
-                statistics: Statistics::default(),
-                projection: None,
-                limit: None,
-                table_partition_cols: vec![],
-            },
-            None,
-            None,
-        )
-        .with_parquet_file_reader_factory(Arc::new(
-            InMemoryParquetFileReaderFactory(Arc::clone(&in_memory_object_store)),
-        ));
-
-        let session_ctx = SessionContext::new();
-
-        let task_ctx = session_ctx.task_ctx();
-        let read = collect(Arc::new(parquet_exec), task_ctx).await.unwrap();
-
-        let expected = vec![
-            "+-----+----+----+",
-            "| c1  | c2 | c3 |",
-            "+-----+----+----+",
-            "| Foo | 1  | 10 |",
-            "|     | 2  | 20 |",
-            "| bar |    |    |",
-            "+-----+----+----+",
-        ];
-
-        assert_batches_sorted_eq!(expected, &read);
-    }
-
-    #[derive(Debug)]
-    struct InMemoryParquetFileReaderFactory(Arc<dyn ObjectStore>);
-
-    impl ParquetFileReaderFactory for InMemoryParquetFileReaderFactory {
-        fn create_reader(
-            &self,
-            partition_index: usize,
-            file_meta: FileMeta,
-            metadata_size_hint: Option<usize>,
-            metrics: &ExecutionPlanMetricsSet,
-        ) -> Result<Box<dyn AsyncFileReader + Send>> {
-            let metadata = file_meta
-                .extensions
-                .as_ref()
-                .expect("has user defined metadata");
-            let metadata = metadata
-                .downcast_ref::<String>()
-                .expect("has string metadata");
-
-            assert_eq!(EXPECTED_USER_DEFINED_METADATA, &metadata[..]);
-
-            let parquet_file_metrics = ParquetFileMetrics::new(
-                partition_index,
-                file_meta.location().as_ref(),
-                metrics,
-            );
-
-            Ok(Box::new(ParquetFileReader {
-                meta: file_meta.object_meta,
-                store: Arc::clone(&self.0),
-                metadata_size_hint,
-                metrics: parquet_file_metrics,
-            }))
-        }
-    }
 
     #[tokio::test]
     async fn evolved_schema() {
